Lab3-Recommender-System/requirements.txt
numpy
scikit-surprise
tqdm

